requirements.txt
flair==0.11.3
inflect==5.6.0
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.5.62
pandas==1.3.2
Pillow==9.1.1
pyspellchecker==0.6.3
requests==2.26.0
sentence_transformers==2.2.0
spellchecker==0.4
torch==1.11.0
torchvision==0.12.0
timm==0.3.2
tqdm==4.64.0
transformers==4.18.0.dev0
yacs==0.1.8